+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Enumeration Types:
 - Add a `JsonEnumeration` interface.
 - Add a `JsonEnumerationType` serialization interface.
 - Add a `JsonStringEnumerationType` implementation.
+- Add a `JsonIntEnumerationType` implementation.
 
 ### Changed